refactoring, options are now in the same file. Same for agents. Agent (excepted keyboard_agent) should not get the object environment
